(UI) refactor PiSDF Popups
