+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/check/PiMMAlgorithmChecker.java
@@ -181,7 +181,7 @@ public class PiMMAlgorithmChecker {
    * @return the ok msg
    */
   public String getOkMsg() {
-    return "Validation of graph " + this.graph.getName() + " raised no error or warning:";
+    return "Validation of graph " + this.graph.getName() + " raised no error or warning.";
   }
 
   /**
